subjectactions - move examples to example_test.go
